+++ src/color_tab.ts
@@ -3,6 +3,7 @@ import * as path from "path";
 import * as os from "os";
 import * as fs from "fs";
 import Storage from "./storage";
+import { Task } from "./taskManager";
 
 interface ColorInfo {
   background: string;
@@ -45,130 +46,149 @@ function getContrastColor(hexColor: string): string {
 }
 
 export function generateCssFile(context: vscode.ExtensionContext): void {
+
   const storage = new Storage(context);
   // set all colors
-  const rulesBasedStyle = vscode.workspace.getConfiguration("tabsColor");
-  const byFileType = rulesBasedStyle.byFileType;
-  const byDirectory = rulesBasedStyle.byDirectory;
-  const activeTab = rulesBasedStyle.activeTab;
   const cssFile = path.join(modulesPath(context), "inject.css");
-  const data = "";
   const tabs = storage.get("tabs") || {};
+  const workspacePath = vscode.workspace.workspaceFolders?.[0]?.uri?.fsPath;
+  const tasks = storage.get(`tasks-${workspacePath}`) || [];
   let style = "";
-  const homeDir = os.homedir() + "/";
-
-  for (const fileType in byFileType) {
-    if (fileType == "filetype") continue;
-    style += `.tab[title$=".${formatTabTitle(fileType)}" i]{background-color:${
-      byFileType[fileType].backgroundColor
-    } !important; opacity:${byFileType[fileType].opacity || "0.6"};}
-          .tab[title$="${formatTabTitle(
-            fileType
-          )}" i] a,.tab[title$="${formatTabTitle(
-      fileType
-    )}" i] .monaco-icon-label:after,.tab[title$="${formatTabTitle(
-      fileType
-    )}" i] .monaco-icon-label:before{color:${
-      byFileType[fileType].fontColor
-    } !important;}`;
-  }
-
-  for (const directory in byDirectory) {
-    if (directory == "my/directory/") continue;
-    const formattedTitle = directory.replace(/\\/g, "\\\\");
-    style += `.tab[title*="${formatTabTitle(
-      formattedTitle
-    )}" i]{background-color:${
-      byDirectory[directory].backgroundColor
-    } !important; opacity: ${byDirectory[directory].opacity || "0.6"};}
-          .tab[title*="${formatTabTitle(
-            formattedTitle
-          )}" i] a,.tab[title*="${formatTabTitle(
-      formattedTitle
-    )}" i] .monaco-icon-label:after,.tab[title*="${formatTabTitle(
-      formattedTitle
-    )}" i] .monaco-icon-label:before{color:${
-      byDirectory[directory].fontColor
-    } !important;}`;
-  }
-
-  let activeSelectors = "";
-  const activeSelectorsArr = [];
+  let activeSelectorsArr: string[] = [];
   const colorsData = storage.get("customColors");
-  for (const i in tabs) {
-    if (colorsData[i]) {
-      const _colorTabs = tabs[i];
+
+  const generateStyleForTabs = (tabs: string[], taskColorDict: any, type: string): string => {
+
+    let style = "";
+    const taskColorData = taskColorDict[type]
+    if (taskColorData) {
       let backgroundSelectors = "";
       let fontColorSelectors = "";
-      const _background = colorsData[i].background;
-      const _fontColor = colorsData[i].color;
-      const _opacity = colorsData[i].opacity || "0.6";
-      const backgroundSelectorsArr = _colorTabs.map(function (a: string) {
+      const _background = taskColorData.background;
+      const _fontColor = taskColorData.color;
+      const _opacity = taskColorData.opacity || "0.6";
+      const backgroundSelectorsArr = tabs.map(function (a: string) {
         return `.tab[title*="${formatTabTitle(a)}" i]`;
       });
       activeSelectorsArr.push(
-        ..._colorTabs.map(function (a: string) {
+        ...tabs.map(function (a: string) {
           return `.tab[title*="${formatTabTitle(a)}" i].active`;
         })
       );
-      const fontColorSelectorsArr = _colorTabs.map(function (a: string) {
-        return `.tab[title*="${formatTabTitle(
-          a
-        )}" i] a,.tab[title="${formatTabTitle(a)}" i] .monaco-icon-label:after,.tab[title*="${formatTabTitle(a)}" i] .monaco-icon-label:before`;
+      const fontColorSelectorsArr = tabs.map(function (a: string) {
+        return `.tab[title*="${formatTabTitle(a)}" i] a,.tab[title="${formatTabTitle(a)}" i] .monaco-icon-label:after,.tab[title*="${formatTabTitle(a)}" i] .monaco-icon-label:before`;
       });
       if (backgroundSelectorsArr.length > 0) {
-        backgroundSelectors =
-          backgroundSelectorsArr.join(",") +
-          `{background-color:${_background} !important; opacity:${_opacity};}`;
+        backgroundSelectors = backgroundSelectorsArr.join(",") + `{background-color:${_background} !important; opacity:${_opacity};}`;
       }
 
       if (fontColorSelectorsArr.length > 0) {
-        fontColorSelectors =
-          fontColorSelectorsArr.join(",") + `{color:${_fontColor} !important;}`;
+        fontColorSelectors = fontColorSelectorsArr.join(",") + `{color:${_fontColor} !important;}`;
       }
       style += backgroundSelectors + fontColorSelectors;
     }
+
     if (activeSelectorsArr.length > 0) {
-      activeSelectors = activeSelectorsArr.join(",") + `{opacity:1;}`;
+      let activeSelectors = activeSelectorsArr.join(",") + `{opacity:1;}`;
+      style += activeSelectors;
     }
-    style += activeSelectors;
-    const dirExists = fs.existsSync(modulesPath(context));
-    if (!dirExists) {
-      const test = fs.mkdirSync(modulesPath(context), { recursive: true });
-      console.log("css", test);
-    }
-    console.log("file", cssFile);
-    if (fs.existsSync(cssFile)) {
-      fs.writeFileSync(cssFile, style);
-    } else {
-      fs.appendFile(cssFile, style, function (err) {
-        if (err) {
-          vscode.window.showInformationMessage(
-            `Could not create a css file. tabscolor won't be able to change your tabs color`
-          );
-          throw err;
+    return style;
+  };
+
+  for (const task of tasks) {
+    if (!task.isComplete){
+      let taskTabs = new Map();
+
+      // Create a map of subtask names to file paths
+      for (const file of task.files) {
+        let fileSubtask = file.subtask || "Task";
+
+        if (!taskTabs.has(fileSubtask)) {
+          taskTabs.set(fileSubtask, []);
         }
-      });
+        taskTabs.get(fileSubtask).push(file.filePath.replace(/\\/g, "\\\\"));
+      }
+
+      // Generate style for each subtask
+      for (const subtask of task.subtasks) {
+        let subtabs = taskTabs.get(subtask.name);
+        if (subtabs) {
+          let subtaskColorData = colorsData[task.id];
+          style += generateStyleForTabs(subtabs, subtaskColorData, subtask.name);
+          
+
+        }
+      }
+
+      // Generate style for files without a subtask
+      let taskFiles = taskTabs.get("Task");
+      if (taskFiles) {
+        let taskColorData = colorsData[task.id];
+        style += generateStyleForTabs(taskFiles, taskColorData, "Task");
+      }
     }
   }
+  const dirExists = fs.existsSync(modulesPath(context));
+  if (!dirExists) {
+    const test = fs.mkdirSync(modulesPath(context), { recursive: true });
+    console.log("css", test);
+  }
+  console.log("file", cssFile);
+  if (fs.existsSync(cssFile)) {
+    fs.writeFileSync(cssFile, style);
+  } else {
+    fs.appendFile(cssFile, style, function (err) {
+      if (err) {
+        vscode.window.showInformationMessage(
+          `Could not create a css file. tabscolor won't be able to change your tabs color`
+        );
+        throw err;
+      }
+    });
+  }
 }
+
+
+
+
 
 export function reloadCss(): void {
   vscode.window.showInformationMessage("---tab updated---");
 }
 
+export function addColor(
+  context: vscode.ExtensionContext,
+  task_id: string,
+  color: string,
+  task_type: string
+) {
+  const storage = new Storage(context);
+  const contrastColor = getContrastColor(color);
+  var colors = { [task_id]: { background: color, color: contrastColor, type: task_type} };
+  storage.addCustomColor(colors);
+}
+
+export function updateColor(
+  context: vscode.ExtensionContext,
+  task_id: string,
+  color: string,
+  task_type: string
+) {
+  const storage = new Storage(context);
+  const contrastColor = getContrastColor(color);
+  var colors = { [task_id]: { background: color, color: contrastColor, type: task_type} };
+  storage.addCustomColor(colors);
+}
+
 export function setColor(
   context: vscode.ExtensionContext,
   task: string,
-  color: string,
-  title: string
+  title: string,
 ): void {
   const storage = new Storage(context);
-  const contrastColor = getContrastColor(color);
+
   if (storage.get("patchedBefore")) {
     if (storage.get("secondActivation")) {
-      var colors = { [task]: { background: color, color: contrastColor } };
-      storage.addCustomColor(colors);
       storage.addTabColor(task, title.replace(/\\/g, "\\\\"));
       generateCssFile(context);
       reloadCss();
@@ -191,7 +211,6 @@ export function unsetColor(
   const storage = new Storage(context);
   if (storage.get("patchedBefore")) {
     if (storage.get("secondActivation")) {
-      storage.removeTabColor(title);
       generateCssFile(context);
       reloadCss();
     } else {