+++ CHANGELOG.md
@@ -1,17 +1,42 @@
 # Change Log
 
+[0.0.11] - 2023-05-28
+
+Added Subtask Beta
+
+Added:
+    - Create Subtasks
+    - Add files to subtasks
+    - Remove files from subtask and subtask themselves
+    - Configuarble shortcuts 
+    - inline Commands
+        - remove task
+        - remove file
+        - remove note
+    - changed shortcuts to use alt instead of shift
+
+Removed:
+    - context Commands:
+        - Create Task
+        - remove task
+        - remove file
+        - remove note
+
+
+NOTE: Subtask is early Beta, Generating a new color for task with subtasks will nto always work and is discouraged 
+
 [0.0.10] - 2023-04-28
 
 Soft lauch with updates based on feedback from intial release
 
 Added:
-    - keyboard shortcut for primary commands
-    - add notes to files
+    - Keyboard shortcut for primary commands
+    - Add notes to files
     - Generate new color
 
 Fixed
-    - note are dynmic base don file changes
-    - file are in task dynamic based on location
+    - Note are dynamic based on file changes
+    - File are in task dynamic based on location
 
 
 [0.0.1] - 2023-04-12